[BREAKING] Remove Num instances for {Const}Expr, Range
They contained bugs in how ranges were added up, also some methods could
not be implemented.

These were mostly just used for integer literals, which can just be
replaced with explicit constructors.

Addition was used for ranges, which I have replaced with `rangeFromSize`
and `simpleRangeToSize`, to convert ranged back and forth with their
sizes. Where addition was used, I have instead converted to a size, done
addition on that and converted back to a range.